refactor: move apiHandler and isAuthorized to Helpers
